refactor(protocol-designer): move step related ui to own branch
As we separate initial state of deck in protocol from data relating to abstract object instance
creation (pipette, liquid, labware), we would like to move purely UI concerns into their own branch
to declutter before further migrations occur. This also stages some low hanging fruit to be moved
into react state to trim down redux in general.
